fixed mistakes

+++ sql/films/create-films.sql
@@ -2,6 +2,6 @@ CREATE TABLE films(
  id serial primary key,
  title varchar(1000) unique,
  genre varchar(1000),
- released Integer(4),
- score Integer(2),
-)
+ released int,
+ score int
+ )
